% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{add_col_totals}
\alias{add_col_totals}
\title{Add column totals for a set of columns in a dataframe}
\usage{
add_col_totals(data, from = 2, to = ncol(data), label = "total",
  lcols = 1)
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
summed by column.}

\item{from}{The number of the column from which column totals are
calculated. The default is 2, assuming one column for row labels. Use 1
if there are no preceding data columns.}

\item{to}{The number of the column to which column totals are calculated.
The default is ncol(data), which means column totals are calculated for
all remaining columms in the dataframe.}

\item{label}{The label for the totals row. The default is "total".}

\item{lcols}{A vector or scalar containing the indices of the columns to
to which to add the total label. These columns must be character vectors.
The default is 1, assuming one column for row labels. Invalid label
columns are silently ignored.}
}
\value{
A tibble containing the input dataframe with an additional row for
  column totals.
}
\description{
Calculates column totals for a set of columns in a dataframe and add them
as a new row at the end of the dataframe. A label for the total row is
added to each of the columns indicated by the \code{label_col} argument.
Label columns must be character vectors. NAs are added to any columns
not summed or specified as label columns.
}
