% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{add_indices}
\alias{add_indices}
\title{Create indices from columns in a dataframe given a baseline row or a vector
of baseline values and add them to the datafame}
\usage{
add_indices(data, from = 2, to = ncol(data), base = 100,
  baserow = 1, basevals = NULL, prefix = "ix_")
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
expressed as indices.}

\item{from}{The number or name of the column from which indices are
calculated. The default is 2, assuming one column for row labels. Use 1
if there are no preceding data columns.}

\item{to}{The number or name of the column to which indices are calculated.
The default is ncol(data), which means indices are calculated across all
remaining columms in the dataframe.}

\item{base}{The baseline value of the created index. The default is 100.
The base must be a positive number greater than zero.}

\item{baserow}{The number of the row to be used as the baseline for
calculating the indices. The default is 1. The baseline row must contain
no zero values. This argument is ignored if the base values are set
explicitly with \code{basevals}.}

\item{basevals}{Specifies explicitly the values to use as the baseline when
calculating the indices. Use this if the baseline values to be used are not
represented in any given row. The default is NULL, which means the
function will use the values in the \code{baserow} instead. The baseline
values must not contain zeros.}

\item{prefix}{A string prefix to add to the column names to identify their
index equivalents. The default is "ix_".}
}
\value{
The input data with additional columns containing indices for
  columns specified by \code{from} and \code{to}.
}
\description{
Creates indices from a set of columns in a dataframe using a given row
number, or vector of values, as the baseline and adds them to the
dataframe.
}
