% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{add_row_percent}
\alias{add_row_percent}
\title{Calculate row percentages for a set of columns in a dataframe and add
them to the dataframe}
\usage{
add_row_percent(data, from = 2, to = ncol(data), na.rm = FALSE,
  prefix = "rp_")
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
expressed as row percentages.}

\item{from}{The number or name of the column from which row percentages are
calculated. The default is 2, assuming one column for row labels.
Use 1 if there are no preceding data columns.}

\item{to}{The number or name of the column to which row percentages are
calculated. The default is ncol(data), which means row percentages are
calculated across all remaining columms in the dataframe.}

\item{na.rm}{A boolean which if TRUE ignores NAs in calculating
percentages. The default value is FALSE.}

\item{prefix}{A string prefix to add to the column names to identify their
percentage equivalents. The default is "rp_"}
}
\value{
The input data with additional columns containing row percentages
  for columns specified by \code{from} and \code{to}.
}
\description{
Calculate row percentages for a set of columns in a dataframe and add
them to the dataframe
}
