% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{add_row_percent_dfr}
\alias{add_row_percent_dfr}
\title{Calculate row percentages for a set of columns in a dataframe and add
them to the dataframe without validation}
\usage{
add_row_percent_dfr(data, from = 2, to = ncol(data), na.rm = FALSE,
  prefix = "rp_")
}
\description{
This is the raw version of \code{add_row_percent}. It should only be used
internally within the package inside functions which provide validation.
}
\keyword{internal}
