% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{add_row_totals}
\alias{add_row_totals}
\title{Add row totals for a set of columns in a dataframe}
\usage{
add_row_totals(data, from = 2, to = ncol(data), label = "total")
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
summed by row.}

\item{from}{The number or name of the column from which row totals are
calculated. The default is 2, assuming one column for row labels. Use 1
if there are no preceding data columns.}

\item{to}{The number or name of the column to which row totals are
calculated. The default is ncol(data), which means row totals are
calculated across all remaining columms in the dataframe.}

\item{label}{The label for the totals column. The default is "total".}
}
\value{
A tibble containing row totals and any preceding columns.
}
\description{
Calculates row totals for a set of columns in a dataframe and add them as a
new column at the end of the dataframe.
}
