% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{add_row_totals_dfr}
\alias{add_row_totals_dfr}
\title{Add row totals for a set of columns in a dataframe without validation}
\usage{
add_row_totals_dfr(data, from = 2, to = ncol(data), label = "total")
}
\description{
This is the raw version of \code{add_row_totals}. It should only be used
internally within the package inside functions which provide validation.
}
\keyword{internal}
