% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltools.R
\docType{package}
\name{cltools}
\alias{cltools}
\title{cltools: A Collection of Data Wrangling Tools for Statistical Researchers}
\description{
The cltools package provides a suit of tools for wrangling tabular data. The
package is particulary helpful for transforming tidy data into summary
tables for presentation. This package is principally designed for use by
researchers in the House of Commons Library but may be useful to anyone
using R for routine data analysis and statistical reporting.
}
