% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjson.R
\name{create_hexjson}
\alias{create_hexjson}
\title{Convert tabular data to hexjson}
\usage{
create_hexjson(data, layout = "odd-r")
}
\arguments{
\item{data}{A dataframe of labels and data to store in each hex.}

\item{layout}{The coordinate layout of the hexsjon. Must be one of:
odd-r, even-r, odd-q, even-q.}
}
\value{
A hexjson string
}
\description{
Converts a dataframe of codes, names and other data to a hexjson string,
adding unique column and row coordinates for each hex. The values in the
first column are used as the key for each hex in the hexjson and therefore
must be unique. This function can be used to produce initial hexjson output
which can then be edited in a hexjson editor.
}
