% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjson.R
\name{create_hexjson_from_csv}
\alias{create_hexjson_from_csv}
\title{Convert a csv of tabular data to hexjson and save it to a file}
\usage{
create_hexjson_from_csv(csv_file, hexjson_file, layout = "odd-r")
}
\arguments{
\item{layout}{The coordinate layout of the hexsjon. Must be one of:
odd-r, even-r, odd-q, even-q.}

\item{csv_name}{The name of a csv file of data to store in each hex.}

\item{hexjson_name}{The name of an output file for the hexjson.}
}
\description{
Converts a csv of codes, names and other data to a hexjson string, adding
unique column and row coordinates for each hex. The values in the first
column are used as the key for each hex in the hexjson and therefore must
be unique. This function can be used to produce initial hexjson output
which can then be edited in a hexjson editor.
}
