% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflators.R
\name{deflate}
\alias{deflate}
\title{Deflate the values in a series by the given deflator}
\usage{
deflate(data, deflator, basepos = length(data))
}
\arguments{
\item{data}{The data series to be deflated as a numeric vector}

\item{deflator}{The deflator series as numeric vector. The deflator series
must be the same length as the data series and must not contain either
NAs or zeros.}

\item{basepos}{The vector index of the baseline for the deflated series:
this equivalent to the period in which all prices will be expressed. The
default value is the index of the last value in the series, which is
assumed to be the most recent period.}
}
\value{
The deflated data series as a numeric vector.
}
\description{
Deflates the values in a series by the given deflator. The data series and
the deflator should be numeric vectors of equal length. The deflator cannot
contain NA or zero values.
}
