% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{get_col_num}
\alias{get_col_num}
\title{Get the column number for a column in a dataframe given its name}
\usage{
get_col_num(data, name)
}
\arguments{
\item{data}{A dataframe.}

\item{name}{A column name as a string.}
}
\value{
The column number for the column with the given name, or zero if
  the column name is not found.
}
\description{
Get the column number for a column in a dataframe given its name
}
\keyword{internal}
