% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{get_col_nums}
\alias{get_col_nums}
\title{Get the column numbers for columns in a dataframe given their names}
\usage{
get_col_nums(data, names)
}
\arguments{
\item{data}{A dataframe.}

\item{name}{A list of column names as strings.}
}
\value{
The column numbers for the columns with the given names, or zero if
  none of the column names are found.
}
\description{
Get the column numbers for columns in a dataframe given their names
}
\keyword{internal}
