% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{get_col_percent}
\alias{get_col_percent}
\title{Calculate column percentages for a set of columns in a dataframe}
\usage{
get_col_percent(data, from = 2, to = NULL)
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
expressed as column percentages.}

\item{from}{The index of the column from which column percentages will
be calculated. The default is 2, assuming one column for row labels.
Use 1 if there are no preceding data columns.}

\item{to}{The index of the column to which column percentages will be
calculated. The default is NULL, which means column percentages will be
calculated across all remaining columms in the dataframe.}
}
\value{
A tibble containing the column percentages and any preceding columns.
}
\description{
Calculates column percentages for a set of colummns in a dataframe and returns
those percentages along with any columns containing row labels or other
preceding data. Use \code{add_col_percent} to append the column
percentages to the input dataframe.
}
