% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{get_col_percent_dfr}
\alias{get_col_percent_dfr}
\title{Calculate column percentages for a set of columns in a dataframe without
validation}
\usage{
get_col_percent_dfr(data, from = 2, to = ncol(data))
}
\description{
This is the raw version of \code{get_col_percent}. It should only be used
internally within the package inside functions which provide validation.
}
\keyword{internal}
