% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{get_col_totals}
\alias{get_col_totals}
\title{Get column totals for a set of columns in a dataframe}
\usage{
get_col_totals(data, from = 2, to = ncol(data))
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
summed by column.}

\item{from}{The number or name of the column from which column totals are
calculated. The default is 2, assuming one column for row labels. Use 1
if there are no preceding data columns.}

\item{to}{The number or name of the column to which column totals are
calculated. The default is ncol(data), which means column totals are
calculated across all remaining columms in the dataframe.}
}
\value{
A numeric vetor of column totals
}
\description{
Calculates column totals for a set of columns in a dataframe and returns
them as a vector. Use \code{add_col_totals} to append the column
percentages to the input dataframe.
}
