% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjson.R
\name{get_hex_string}
\alias{get_hex_string}
\title{Create a hexjson string for a hex from a dataframe row}
\usage{
get_hex_string(...)
}
\arguments{
\item{...}{A row of data from a dataframe.}
}
\value{
The hex part of a hexjson string.
}
\description{
Creates a hexjson string for an individual hex from a dataframe row. The
column names are used as the property names and the vales are used as the
property values. The value in the first column is used as the key for the
hex in the hexjson.
}
\keyword{internal}
