% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{get_index}
\alias{get_index}
\title{Create a statistical index from a numeric vector given a baseline vector
index or a baseline value}
\usage{
get_index(v, base = 100, basepos = 1, baseval = NULL)
}
\arguments{
\item{v}{A numeric vector containing values to be expressed as an index.}

\item{base}{The baseline value of the created index. The default is 100.
The base must be a positive number greater than zero.}

\item{basepos}{The vector index of the value to be used as the baseline for
calculating the index. The default is 1. The value at basepos must not be
zero. This argument is ignored if the base value is set explcitly with
\code{baseval}.}

\item{baseval}{Specifies explicitly the value to use as the baseline when
calculating the index. Use this if the baseline value to be used is not
represented in the vector. The default is NULL, which means the function
will use the values in the \code{baserow} instead. The baseline value
must not be zero.}
}
\value{
A vector containing the created index.
}
\description{
Creates a statistical index from a numeric vector given a baseline
vector index or a baseline value.
}
