% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{get_indices_dfr}
\alias{get_indices_dfr}
\title{Create indices from columns in a dataframe given a baseline row or a vector
of baseline values without validation}
\usage{
get_indices_dfr(data, from = 2, to = ncol(data), start = 100,
  baserow = 1, basevals = NULL)
}
\description{
This is the raw version of \code{get_indices}. It should only be used
internally within the package inside functions which provide validation.
}
\keyword{internal}
