% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflators.R
\name{get_real_dfr}
\alias{get_real_dfr}
\title{Create real terms series from columns in a dataframe given a deflator
series and a baseline row without validation}
\usage{
get_real_dfr(data, deflator, from = 2, to = ncol(data),
  baserow = nrow(data))
}
\description{
This is the raw version of \code{get_real}. It should only be used
internally within the package inside functions which provide validation.
}
\keyword{internal}
