% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{get_row_totals_dfr}
\alias{get_row_totals_dfr}
\title{Get row totals for a set of columns in a dataframe without validation}
\usage{
get_row_totals_dfr(data, from = 2, to = ncol(data), na.rm = FALSE)
}
\description{
This is the raw version of \code{get_row_totals}. It should only be used
internally within the package inside functions which provide validation.
}
\keyword{internal}
