% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{run_dfr_func}
\alias{run_dfr_func}
\title{Run a dataframe range function with validation}
\usage{
run_dfr_func(f, data, from, to, ...)
}
\arguments{
\item{f}{The dataframe range function.}

\item{from}{The start column: either an index or name.}

\item{to}{The end column: either and index or name.}
}
\value{
The retrun value of the function.
}
\description{
A dataframe range function is a generic name for a set of functions in this
package with a common signature. Dataframe range funtions take a dataframe
along with \code{from} and \code{to} arguments, which specify a range of
columns over which operations will be performed.
}
\details{
This function provides standard validation for the arguments that are
common to these functions. It takes a dataframe range function, checks the
standard inputs are valid, runs the function and returns any values.
}
\keyword{internal}
