% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{valid_args}
\alias{valid_args}
\title{Validate the arguments to a dataframe range function are the right type
and size.}
\usage{
valid_args(data, from, to)
}
\arguments{
\item{data}{The input data.}

\item{from}{The number or name of the column from which data is processed.}

\item{to}{The number or name of the column to which data is processed.}
}
\value{
TRUE if no errors are thrown.
}
\description{
Validate the arguments to a dataframe range function are the right type
and size.
}
\keyword{internal}
