% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{valid_columns}
\alias{valid_columns}
\title{Validate column arguments for row and column percentages and totals}
\usage{
valid_columns(data, from, to)
}
\arguments{
\item{data}{A dataframe containing columns of numerical data for analysis.}

\item{from}{The index of the column from which data is processed.}

\item{to}{The index of the column to which data is processed.}
}
\description{
Validate column arguments for row and column percentages and totals
}
\keyword{internal}
