% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cltools.R
\docType{package}
\name{cltools}
\alias{cltools}
\alias{cltools-package}
\title{cltools: A Collection of Data Wrangling Tools for Statistical Researchers}
\description{
cltools: A Collection of Data Wrangling Tools for Statistical Researchers
}
