% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{get_col_percent}
\alias{get_col_percent}
\title{Calculate column percentages for a set of columns in a dataframe}
\usage{
get_col_percent(data, from = 2, to = ncol(data), na.rm = FALSE)
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
expressed as column percentages.}

\item{from}{The number or name of the column from which column percentages
are calculated. The default is 2, assuming one column for row labels.
Use 1 if there are no preceding data columns.}

\item{to}{The number or name of the column to which column percentages are
calculated. The default is ncol(data), which means column percentages are
calculated across all remaining columms in the dataframe.}

\item{na.rm}{A boolean which if TRUE ignores NAs in calculating
percentages. The default value is FALSE.}
}
\value{
A tibble containing column percentages and any preceding columns.
}
\description{
Calculates column percentages for a set of colummns in a dataframe and
returns those percentages along with any preceding columns containing row
labels or other data. Use \code{add_col_percent} to append the column
percentages to the input dataframe.
}
