% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{get_row_percent}
\alias{get_row_percent}
\title{Calculate row percentages for a set of columns in a dataframe}
\usage{
get_row_percent(data, from = 2, to = ncol(data), na.rm = FALSE)
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
expressed as row percentages.}

\item{from}{The number or name of the column from which row percentages are
calculated. The default is 2, assuming one column for row labels.
Use 1 if there are no preceding data columns.}

\item{to}{The number or name of the column to which row percentages are
calculated. The default is ncol(data), which means row percentages are
calculated across all remaining columms in the dataframe.}

\item{na.rm}{A boolean which if TRUE ignores NAs in calculating
percentages. The default value is FALSE.}
}
\value{
A tibble containing row percentages and any preceding columns.
}
\description{
Calculates row percentages for a set of columns in a dataframe and returns
those percentages along with any preceding columns containing row labels or
other data. Use \code{add_row_percent} to append the row percentages to the
input dataframe.
}
