% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{get_row_totals}
\alias{get_row_totals}
\title{Get row totals for a set of columns in a dataframe}
\usage{
get_row_totals(data, from = 2, to = ncol(data))
}
\arguments{
\item{data}{A dataframe containing columns of numerical data to be
summed by row.}

\item{from}{The number of the column from which row totals are calculated.
The default is 2, assuming one column for row labels. Use 1 if there are
no preceding data columns.}

\item{to}{The number of the column to which row totals are calculated. The
default is ncol(data), which means row totals are calculated across all
remaining columms in the dataframe.}
}
\value{
A numeric vetor of row totals
}
\description{
Calculates row totals for a set of columns in a dataframe and returns them
as a vector. Use \code{add_row_totals} to append the row percentages to the
input dataframe.
}
