% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{valid_columns}
\alias{valid_columns}
\title{Validate column arguments for row percentages, totals and indices}
\usage{
valid_columns(data, from, to)
}
\arguments{
\item{data}{A dataframe containing columns of numerical data for analysis.}

\item{from}{The number of the column from which data is processed.}

\item{to}{The number of the column to which data is processed.}
}
\value{
TRUE if no errors are thrown.
}
\description{
Validate column arguments for row percentages, totals and indices
}
\keyword{internal}
