% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_df}
\alias{valid_df}
\title{Validate input data is a dataframe}
\usage{
valid_df(data)
}
\arguments{
\item{data}{The input data.}
}
\description{
Validate input data is a dataframe
}
\keyword{internal}
